+++ lib/http.js
@@ -4,177 +4,130 @@
 
 const http = require('http')
 const https = require('https')
+const { stringify } = require('querystring')
+const { env } = require('./init')
 const { debuglog } = require('util')
-const { createResultsParser } = require('./json')
-const { del, put } = require('./level')
 
 const debug = debuglog('fanboy')
 
 exports.request = request
-exports.createResultsParser = createResultsParser
+exports.createPath = createPath
 
-// Request lookup iTunes ID or search for term. Optional `keys` are used as
-// fallback.
-//
-// - term String iTunes ID or search term.
-// - keys [String] Array of cached keys (optional).
-// - delegate Transform The surface transform stream.
-// - cb (er) => {} A standard callback.
-function request (term, keys, delegate, cb) {
-  if (typeof keys === 'function') {
-    cb = keys
-    keys = null
+const verbs = { '/search': 'term', '/lookup': 'id' }
+
+function decorate (obj, path, term) {
+  obj[verbs[path]] = term
+
+  return obj
+}
+
+/**
+ * Returns a path constructed for the iTunes API.
+ *
+ * @throws Throws 'unknown path' if the local `verbs` have no match.
+ */
+function createPath (path, term, { media = 'all', country = 'us', attribute } = {}) {
+  if (!verbs.hasOwnProperty(path)) {
+    throw new Error('unknown path')
   }
 
-  const skip = () => {
-    const y = delegate.cache.has(term)
+  const obj = (() => {
+    if (path === '/lookup') return Object.create(null)
 
-    if (y) debug('skipping: %s', term)
-
-    return y
-  }
-
-  if (skip()) {
-    return cb()
-  }
-
-  const opts = delegate.reqOpts(term)
-  debug(opts)
-
-  const fallback = () => {
-    debug('falling back')
-    if (skip()) return cb()
-    if (keys) {
-      return delegate.resultsForKeys(keys, cb)
+    const o = {
+      media: media,
+      country: country
     }
 
-    delegate.keysForTerm(term, (er, keys) => {
-      if (er) {
-        return cb(er.notFound ? null : er)
-      }
-      delegate.resultsForKeys(keys, cb)
-    })
+    if (attribute) o.attribute = attribute
+
+    return o
+  })()
+
+  // Note: URL encoding replaces spaces with the plus (+) character and all
+  // characters except the following are encoded: letters, numbers, periods
+  // (.), dashes (-), underscores (_), and asterisks (*).
+
+  const q = stringify(decorate(obj, path, term))
+
+  return [path, q].join('?')
+}
+
+function ReqOpts ({
+  hostname = 'localhost',
+  keepAlive = true,
+  port = 8080,
+  method = 'GET',
+  path = '/'
+}) {
+  this.hostname = hostname
+  this.keepAlive = keepAlive
+  this.method = method
+  this.path = path
+  this.port = port
+}
+
+/**
+ * Issues an HTTP request at `path` ('/search' or '/lookup') to query `term`,
+ * passing resulting error and response into `cb` once the response body gets
+ * available to be consumed. Upon receiving an error, the callback should
+ * dismiss the readable response stream without adding listeners to it.
+ *
+ * Response handling lies in the responsibility of the callback. Pass `opts`
+ * for overriding global environment settings during testing.
+ */
+function request ({ path, term, responseHandler }, opts = env()) {
+  const onabort = () => {
+    const err = new Error('fanboy: request aborted')
+
+    done(err)
+  }
+
+  const onerror = (error) => {
+    const err = Object.assign(new Error(), error)
+    err.message = 'fanboy: ' + error.message
+
+    done(err)
   }
 
   const onresponse = (res) => {
-    const { statusCode, headers } = res
+    // TODO: Inspect headers: who are we talking to?
 
-    debug('HTTP ( %s, %o )', statusCode, headers)
-
-    if (statusCode !== 200) {
-      const er = new Error('fanboy: unexpected response ' + statusCode)
-      er.statusCode = statusCode
-      er.headers = res.headers
-
-      Object.assign(er, opts)
-
-      return done(er)
-    }
-
-    // Parsing
-
-    const parser = createResultsParser(res)
-    const results = []
-
-    function ondrain () {
-      debug('drain event received')
-      parser.resume()
-    }
-
-    const ondata = (obj) => {
-      const result = delegate.result(obj)
-
-      if (result) {
-        results.push(result)
-
-        const chunk = JSON.stringify(result)
-
-        if (!delegate.use(chunk)) {
-          debug('waiting for drain while readable flowing: %s', delegate.readableFlowing)
-          delegate.once('drain', ondrain)
-          parser.pause()
-        }
-      }
-    }
-
-    // If faulty, results are not trustworthy.
-    let faulty = false
-
-    const onend = () => {
-      if (results.length) {
-        put(delegate.db, term, results, (er) => {
-          parsed(er)
-        })
-      } else if (!faulty) {
-        del(delegate.db, term, (er) => {
-          delegate.cache.set(term, true)
-          parsed(er)
-        })
-      } else {
-        parsed()
-      }
-    }
-
-    function onerror (error) {
-      faulty = true
-      const er = new Error('fanboy: parse error: ' + error.message)
-
-      parsed(er)
-    }
-
-    const parsed = (er) => {
-      delegate.removeListener('drain', ondrain)
-
-      parser.removeListener('data', ondata)
-      parser.removeListener('end', onend)
-      parser.removeListener('error', onerror)
-
-      done(er)
-    }
-
-    parser.on('data', ondata)
-    parser.once('end', onend)
-    parser.once('error', onerror)
+    done(null, res)
   }
 
-  // Requesting
+  const p = createPath(path, term, env())
+  const reqOpts = new ReqOpts(Object.assign({ path: p }, opts))
+
+  debug('issueing request: %o', reqOpts)
 
   const mod = opts.port === 443 ? https : http
-  const req = mod.request(opts, onresponse)
+  const req = mod.request(reqOpts, onresponse)
 
-  const done = (er) => {
-    req.removeListener('aborted', onaborted)
+  const ontimeout = () => {
+    debug('aborting request: socket has been idle')
+
+    // Assuming we will redundantly receive ECONNRESET, but that might be
+    // a Nock quirk. Needs investigation.
+
+    req.removeListener('abort', onabort)
+    req.abort()
+  }
+
+  // Before leaving this scope, call done with error and response.
+  function done (err, res) {
+    req.removeListener('abort', onabort)
     req.removeListener('error', onerror)
     req.removeListener('response', onresponse)
+    req.removeListener('timeout', ontimeout)
 
-    if (er) {
-      if (keys) {
-        er.message = 'fanboy: falling back on cache'
-
-        delegate.emit('error', er)
-
-        return fallback()
-      }
-    }
-
-    if (cb) cb(er)
+    if (responseHandler) responseHandler(err, res)
   }
 
-  function onaborted () {
-    const er = new Error('fanboy: request aborted')
-
-    done(er)
-  }
-
-  function onerror (error) {
-    const er = Object.assign(new Error(), error)
-    er.message = 'fanboy: ' + error.message
-
-    done(er)
-  }
-
-  req.once('aborted', onaborted)
+  req.once('abort', onabort)
   req.once('error', onerror)
+  req.once('timeout', ontimeout)
+  req.setTimeout(3000)
 
-  req.end()
+  return req
 }
