+++ lib/lookup.js
@@ -1,23 +1,23 @@
 'use strict'
 
-// search.js - search iTunes
+// lookup.js -  lookup specific items
 
-const { keysForTerm, createLevelRandom, writeResults, guid } = require('../level')
-const { Readable, Writable, pipeline, Transform } = require('stream')
-const { jsonResultsParser } = require('../json')
+const { writeResults, resultForID, guid } = require('./level')
+const { Writable, pipeline, Transform } = require('stream')
+const { jsonResultsParser } = require('./json')
 const { request } = require('./http')
 const { debuglog } = require('util')
 
 const debug = debuglog('fanboy')
 
-exports.search = search
+exports.lookup = lookup
 
 /**
  * Issues an HTTP request storing the result. When done, the callback receives
- * an error if something went wrong and hopefully an Array of results.
+ * an error if something went wrong and hopefully the matching item.
  */
-function requestSearch (term, { db, cache }, cb) {
-  debug('requesting search: %s', term)
+function requestLookup (iTunesId, { db, cache }, cb) {
+  debug('requesting lookup: %s', iTunesId)
 
   const responseHandler = (err, res) => {
     if (err) {
@@ -29,7 +29,7 @@ function requestSearch (term, { db, cache }, cb) {
       return cb(new Error(`unexpected HTTP status code: ${res.statusCode}`))
     }
 
-    debug('creating search request pipeline')
+    debug('creating lookup request pipeline')
 
     const results = []
 
@@ -70,68 +70,52 @@ function requestSearch (term, { db, cache }, cb) {
         //   return cb(new Error('terminated while the message was still being sent'))
         // }
 
+        debug(chunksCount)
+
         if (chunksCount === 0) {
           return cb(new Error('nothing to read'))
         }
 
-        writeResults({ db, cache, term, results }, cb)
+        writeResults({ db, cache, term: iTunesId, results }, (err, items) => {
+          cb(err, items ? items[0] : null)
+        })
       }
     )
   }
 
   request({
-    path: '/search',
-    term: term,
+    path: '/lookup',
+    term: iTunesId,
     responseHandler: responseHandler
   }).end()
 }
 
-/**
- * Reads values for `keys` from `db`.
- */
-function read (db, keys, cb) {
-  const results = []
-
-  pipeline(
-    new Readable({
-      read (size) {
-        const key = keys.shift()
-        const chunk = key || null
-        debug('pushing key: %s', chunk)
-        this.push(chunk)
-      },
-      objectMode: true
-    }),
-    createLevelRandom(db),
-    new Writable({
-      write (chunk, enc, cb) {
-        results.push(chunk)
-        cb()
-      }
-    }),
-    err => {
-      debug(results)
-      cb(err, results)
-    }
-  )
+function isGUID (value) {
+  return parseInt(value, 10)
 }
 
 /**
- * Search items matching term. The two code paths are:
+ * Lookup items matching guids. The two code paths are:
  *
- * - HTTP request and write to database
+ * - HTTP request and update database
  * - Read values from database
  */
-function search (term, { db, ttl, cache }, cb) {
-  keysForTerm(db, term, ttl, (err, keys) => {
+function lookup (iTunesId, { db, ttl, cache }, cb) {
+  if (!isGUID(iTunesId)) {
+    return cb(new Error('invalid iTunes ID'))
+  }
+
+  resultForID(db, iTunesId, (err, item) => {
     if (err) {
       if (err.notFound) {
-        requestSearch(term, { db, cache }, cb)
+        requestLookup(iTunesId, { db, cache }, cb)
       } else {
         cb(err)
       }
+    } else if (item !== undefined) {
+      cb(null, item)
     } else {
-      read(db, keys, cb)
+      cb(new Error('unexpected database problem'))
     }
   })
 }
