+++ lib/search.js
@@ -1,23 +1,23 @@
 'use strict'
 
-// lookup.js -  lookup specific items
+// search.js - search iTunes
 
-const { writeResults, resultForID, guid } = require('../level')
-const { Writable, pipeline, Transform } = require('stream')
-const { jsonResultsParser } = require('../json')
+const { keysForTerm, createLevelRandom, writeResults, guid } = require('./level')
+const { Readable, Writable, pipeline, Transform } = require('stream')
+const { jsonResultsParser } = require('./json')
 const { request } = require('./http')
 const { debuglog } = require('util')
 
 const debug = debuglog('fanboy')
 
-exports.lookup = lookup
+exports.search = search
 
 /**
  * Issues an HTTP request storing the result. When done, the callback receives
- * an error if something went wrong and hopefully the matching item.
+ * an error if something went wrong and hopefully an Array of results.
  */
-function requestLookup (iTunesId, { db, cache }, cb) {
-  debug('requesting lookup: %s', iTunesId)
+function requestSearch (term, { db, cache }, cb) {
+  debug('requesting search: %s', term)
 
   const responseHandler = (err, res) => {
     if (err) {
@@ -29,7 +29,7 @@ function requestLookup (iTunesId, { db, cache }, cb) {
       return cb(new Error(`unexpected HTTP status code: ${res.statusCode}`))
     }
 
-    debug('creating lookup request pipeline')
+    debug('creating search request pipeline')
 
     const results = []
 
@@ -70,52 +70,68 @@ function requestLookup (iTunesId, { db, cache }, cb) {
         //   return cb(new Error('terminated while the message was still being sent'))
         // }
 
-        debug(chunksCount)
-
         if (chunksCount === 0) {
           return cb(new Error('nothing to read'))
         }
 
-        writeResults({ db, cache, term: iTunesId, results }, (err, items) => {
-          cb(err, items ? items[0] : null)
-        })
+        writeResults({ db, cache, term, results }, cb)
       }
     )
   }
 
   request({
-    path: '/lookup',
-    term: iTunesId,
+    path: '/search',
+    term: term,
     responseHandler: responseHandler
   }).end()
 }
 
-function isGUID (value) {
-  return parseInt(value, 10)
+/**
+ * Reads values for `keys` from `db`.
+ */
+function read (db, keys, cb) {
+  const results = []
+
+  pipeline(
+    new Readable({
+      read (size) {
+        const key = keys.shift()
+        const chunk = key || null
+        debug('pushing key: %s', chunk)
+        this.push(chunk)
+      },
+      objectMode: true
+    }),
+    createLevelRandom(db),
+    new Writable({
+      write (chunk, enc, cb) {
+        results.push(chunk)
+        cb()
+      }
+    }),
+    err => {
+      debug(results)
+      cb(err, results)
+    }
+  )
 }
 
 /**
- * Lookup items matching guids. The two code paths are:
+ * Search items matching term. The two code paths are:
  *
- * - HTTP request and update database
+ * - HTTP request and write to database
  * - Read values from database
  */
-function lookup (iTunesId, { db, ttl, cache }, cb) {
-  if (!isGUID(iTunesId)) {
-    return cb(new Error('invalid iTunes ID'))
-  }
-
-  resultForID(db, iTunesId, (err, item) => {
+function search (term, { db, ttl, cache }, cb) {
+  keysForTerm(db, term, ttl, (err, keys) => {
     if (err) {
       if (err.notFound) {
-        requestLookup(iTunesId, { db, cache }, cb)
+        requestSearch(term, { db, cache }, cb)
       } else {
         cb(err)
       }
-    } else if (item !== undefined) {
-      cb(null, item)
     } else {
-      cb(new Error('unexpected database problem'))
+      read(db, keys, cb)
     }
   })
 }
