+++ lib/suggest.js
@@ -2,7 +2,7 @@
 
 // suggest.js -  get search suggestions from iTunes
 
-const { keyStream, termFromKey } = require('../level')
+const { keyStream, termFromKey } = require('./level')
 const { debuglog } = require('util')
 const { Writable, pipeline } = require('stream')
